Refactor: move combinations function to g2.py and update g3.py to use it
